requirements.txt
ipython==7.10.1
numpy==1.17.4
Pillow==6.2.1
tqdm==4.40.1
matplotlib==3.1.2

